requirements.txt
discord.py==1.7.3
requests==2.25.1
flask==2.0.1
flask-discord==0.1.5
python-dotenv==0.19.2
ftplib==3.4.0
sqlite3==3.35.5
